# [Scaling Writers] Close Idle Writers