--- a/core/trino-main/src/main/java/io/trino/operator/TableWriterOperator.java
+++ b/core/trino-main/src/main/java/io/trino/operator/TableWriterOperator.java
@@ -20,7 +20,9 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.primitives.Ints;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
+import io.airlift.log.Logger;
 import io.airlift.slice.Slice;
+import io.airlift.units.DataSize;
 import io.airlift.units.Duration;
 import io.trino.Session;
 import io.trino.memory.context.LocalMemoryContext;
@@ -42,6 +44,7 @@ import io.trino.util.AutoCloseableCloser;
 
 import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Supplier;
@@ -53,6 +56,8 @@ import static com.google.common.util.concurrent.Futures.allAsList;
 import static com.google.common.util.concurrent.MoreExecutors.directExecutor;
 import static io.airlift.concurrent.MoreFutures.getFutureValue;
 import static io.airlift.concurrent.MoreFutures.toListenableFuture;
+import static io.trino.SystemSessionProperties.getCloseIdleWritersTriggerDuration;
+import static io.trino.SystemSessionProperties.getIdleWriterMinDataSizeThreshold;
 import static io.trino.SystemSessionProperties.isStatisticsCpuTimerEnabled;
 import static io.trino.spi.type.BigintType.BIGINT;
 import static io.trino.spi.type.VarbinaryType.VARBINARY;
@@ -64,6 +69,7 @@ import static java.util.concurrent.TimeUnit.NANOSECONDS;
 public class TableWriterOperator
         implements Operator
 {
+    private static final Logger LOG = Logger.get(TableWriterOperator.class);
     public static final int ROW_COUNT_CHANNEL = 0;
     public static final int FRAGMENT_CHANNEL = 1;
     public static final int STATS_START_CHANNEL = 2;
@@ -111,10 +117,22 @@ public class TableWriterOperator
         public Operator createOperator(DriverContext driverContext)
         {
             checkState(!closed, "Factory is already closed");
+            // Driver should call getOutput() periodically on TableWriterOperator to close idle writers which will essentially
+            // decrease the memory usage even if no pages were added to that writer thread.
+            if (getCloseIdleWritersTriggerDuration(session).toMillis() > 0) {
+                driverContext.setBlockedTimeout(getCloseIdleWritersTriggerDuration(session));
+            }
             OperatorContext context = driverContext.addOperatorContext(operatorId, planNodeId, TableWriterOperator.class.getSimpleName());
             Operator statisticsAggregationOperator = statisticsAggregationOperatorFactory.createOperator(driverContext);
             boolean statisticsCpuTimerEnabled = !(statisticsAggregationOperator instanceof DevNullOperator) && isStatisticsCpuTimerEnabled(session);
-            return new TableWriterOperator(context, createPageSink(driverContext), columnChannels, statisticsAggregationOperator, types, statisticsCpuTimerEnabled);
+            return new TableWriterOperator(
+                    context,
+                    createPageSink(driverContext),
+                    columnChannels,
+                    statisticsAggregationOperator,
+                    types,
+                    statisticsCpuTimerEnabled,
+                    getIdleWriterMinDataSizeThreshold(session));
         }
 
         private ConnectorPageSink createPageSink(DriverContext driverContext)
@@ -159,6 +177,7 @@ public class TableWriterOperator
     private final AtomicLong pageSinkPeakMemoryUsage = new AtomicLong();
     private final Operator statisticAggregationOperator;
     private final List<Type> types;
+    private final DataSize idleWriterMinDataSizeThreshold;
 
     private ListenableFuture<Void> blocked = NOT_BLOCKED;
     private CompletableFuture<Collection<Slice>> finishFuture;
@@ -170,8 +189,10 @@ public class TableWriterOperator
 
     private final OperationTiming statisticsTiming = new OperationTiming();
     private final boolean statisticsCpuTimerEnabled;
-
     private final Supplier<TableWriterInfo> tableWriterInfoSupplier;
+    // This records the last physical written data size when connector closeIdleWriters is triggered.
+    private long lastPhysicalWrittenDataSize;
+    private boolean newPagesAdded;
 
     public TableWriterOperator(
             OperatorContext operatorContext,
@@ -179,7 +200,8 @@ public class TableWriterOperator
             List<Integer> columnChannels,
             Operator statisticAggregationOperator,
             List<Type> types,
-            boolean statisticsCpuTimerEnabled)
+            boolean statisticsCpuTimerEnabled,
+            DataSize idleWriterMinDataSizeThreshold)
     {
         this.operatorContext = requireNonNull(operatorContext, "operatorContext is null");
         this.pageSinkMemoryContext = operatorContext.newLocalUserMemoryContext(TableWriterOperator.class.getSimpleName());
@@ -188,6 +210,7 @@ public class TableWriterOperator
         this.statisticAggregationOperator = requireNonNull(statisticAggregationOperator, "statisticAggregationOperator is null");
         this.types = ImmutableList.copyOf(requireNonNull(types, "types is null"));
         this.statisticsCpuTimerEnabled = statisticsCpuTimerEnabled;
+        this.idleWriterMinDataSizeThreshold = requireNonNull(idleWriterMinDataSizeThreshold, "idleWriterMinDataSizeThreshold is null");
         this.tableWriterInfoSupplier = createTableWriterInfoSupplier(pageSinkPeakMemoryUsage, statisticsTiming, pageSink);
         this.operatorContext.setInfoSupplier(tableWriterInfoSupplier);
     }
@@ -259,14 +282,20 @@ public class TableWriterOperator
         rowCount += page.getPositionCount();
         updateWrittenBytes();
         operatorContext.recordWriterInputDataSize(page.getSizeInBytes());
+        newPagesAdded = true;
     }
 
     @Override
     public Page getOutput()
     {
-        if (!blocked.isDone()) {
+        tryClosingIdleWriters();
+        // This method could be called even when new pages have not been added. In that case, we don't have to
+        // try to get the output from the aggregation operator. It could be expensive since getOutput() is
+        // called quite frequently.
+        if (!(blocked.isDone() && (newPagesAdded || state != State.RUNNING))) {
             return null;
         }
+        newPagesAdded = false;
 
         if (!statisticAggregationOperator.isFinished()) {
             OperationTimer timer = new OperationTimer(statisticsCpuTimerEnabled);
@@ -363,6 +392,24 @@ public class TableWriterOperator
         long current = pageSink.getCompletedBytes();
         operatorContext.recordPhysicalWrittenData(current - writtenBytes);
         writtenBytes = current;
+    }
+
+    private void tryClosingIdleWriters()
+    {
+        long physicalWrittenDataSize = getTaskContext().getPhysicalWrittenDataSize();
+        Optional<Integer> writerCount = getTaskContext().getMaxWriterCount();
+        if (writerCount.isEmpty() || physicalWrittenDataSize - lastPhysicalWrittenDataSize <= idleWriterMinDataSizeThreshold.toBytes() * writerCount.get()) {
+            return;
+        }
+        pageSink.closeIdleWriters();
+        updateMemoryUsage();
+        updateWrittenBytes();
+        lastPhysicalWrittenDataSize = physicalWrittenDataSize;
+    }
+
+    private TaskContext getTaskContext()
+    {
+        return operatorContext.getDriverContext().getPipelineContext().getTaskContext();
     }
 
     private void updateMemoryUsage()