--- a/lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/classloader/ClassLoaderSafeConnectorPageSink.java
+++ b/lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/classloader/ClassLoaderSafeConnectorPageSink.java
@@ -70,6 +70,14 @@ public class ClassLoaderSafeConnectorPageSink
     }
 
     @Override
+    public void closeIdleWriters()
+    {
+        try (ThreadContextClassLoader ignored = new ThreadContextClassLoader(classLoader)) {
+            delegate.closeIdleWriters();
+        }
+    }
+
+    @Override
     public CompletableFuture<Collection<Slice>> finish()
     {
         try (ThreadContextClassLoader ignored = new ThreadContextClassLoader(classLoader)) {