… core/trino-main/src/test/java/io/trino/execution/MockRemoteTaskFactory.java
… core/trino-main/src/test/java/io/trino/execution/TestSqlTask.java
… core/trino-main/src/test/java/io/trino/execution/TestSqlTaskExecution.java
… core/trino-main/src/test/java/io/trino/execution/TestTaskManagerConfig.java
… core/trino-main/src/test/java/io/trino/memory/TestMemoryPools.java
… core/trino-main/src/test/java/io/trino/memory/TestMemoryTracking.java
… core/trino-main/src/test/java/io/trino/operator/GroupByHashYieldAssertion.java
… core/trino-main/src/test/java/io/trino/operator/TestDriver.java
… core/trino-main/src/test/java/io/trino/operator/TestingOperatorContext.java
… plugin/trino-delta-lake/src/main/java/io/trino/plugin/deltalake/AbstractDeltaLakePageSink.java
… plugin/trino-delta-lake/src/main/java/io/trino/plugin/deltalake/DeltaLakeConfig.java
… plugin/trino-delta-lake/src/main/java/io/trino/plugin/deltalake/DeltaLakeSessionProperties.java
… plugin/trino-delta-lake/src/test/java/io/trino/plugin/deltalake/TestCloseIdleWriters.java
… plugin/trino-delta-lake/src/test/java/io/trino/plugin/deltalake/TestDeltaLakeConfig.java
… plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveConfig.java
… plugin/trino-hive/src/main/java/io/trino/plugin/hive/HivePageSink.java
… plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveSessionProperties.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestCloseIdleWriters.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveConfig.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHivePageSink.java
… plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/IcebergConfig.java
… plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/IcebergPageSink.java
… plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/IcebergSessionProperties.java
… plugin/trino-iceberg/src/test/java/io/trino/plugin/iceberg/TestCloseIdleWriters.java
… plugin/trino-iceberg/src/test/java/io/trino/plugin/iceberg/TestIcebergConfig.java